Move Core domain models to have vertical slices
