+++ backend/Api/Organisation/OrganisationHolidayExtensions.cs
@@ -1,4 +1,5 @@
 using Core.DomainModels;
+using Core.Organization;
 using PublicHoliday;
 
 namespace Api.Organisation;
