+++ backend/Api/Projects/ProjectController.cs
@@ -1,7 +1,8 @@
 using Api.Common;
 using Api.StaffingController;
 using Core.DomainModels;
-using Core.IRepositories;
+using Core.Engagement;
+using Core.Organization;
 using Infrastructure.DatabaseContext;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
