+++ backend/Core/DomainModels/Absence.cs
@@ -9,5 +9,5 @@ public class Absence
 
     public required string Name { get; set; }
     public required bool ExcludeFromBillRate { get; set; } = false;
-    public required Organization Organization { get; set; }
+    public required Organization.Organization Organization { get; set; }
 }
