+++ backend/Core/Organization/IOrganisationRepository.cs
@@ -1,6 +1,4 @@
-using Core.DomainModels;
-
-namespace Core.IRepositories;
+namespace Core.Organization;
 
 public interface IOrganisationRepository
 {
