+++ backend/Infrastructure/Repositories/EngagementDbRepository.cs
@@ -1,5 +1,5 @@
 using Core.DomainModels;
-using Core.IRepositories;
+using Core.Engagement;
 using Infrastructure.DatabaseContext;
 using Microsoft.EntityFrameworkCore;
